refactor: Update SideBarDrawer component to use LogoutUser utility function
