+++ platform/src/common/components/TopBar/index.jsx
@@ -2,23 +2,19 @@ import { useState, useEffect } from 'react';
 import SearchMdIcon from '@/icons/Common/search_md.svg';
 import TopBarItem from './TopBarItem';
 import { useRouter } from 'next/router';
-import { resetStore } from '@/lib/store/services/account/LoginSlice';
-import { resetChartStore } from '@/lib/store/services/charts/ChartSlice';
 import { useSelector, useDispatch } from 'react-redux';
 import MenuBarIcon from '@/icons/menu_bar';
 import AirqoLogo from '@/icons/airqo_logo.svg';
 import ExpandIcon from '@/icons/SideBar/expand.svg';
-import { resetAllTasks } from '@/lib/store/services/checklists/CheckList';
-import { updateUserChecklists, resetChecklist } from '@/lib/store/services/checklists/CheckData';
 import Spinner from '@/components/Spinner';
 import SettingsIcon from '@/icons/SideBar/SettingsIcon';
 import UserIcon from '@/icons/Topbar/userIcon';
-import { clearIndividualPreferences } from '@/lib/store/services/account/UserDefaultsSlice';
 import {
   toggleSidebar,
   setToggleDrawer,
   setSidebar,
 } from '@/lib/store/services/sideBar/SideBarSlice';
+import LogoutUser from '@/core/utils/LogoutUser';
 
 const TopBar = ({ topbarTitle, noBorderBottom }) => {
   // check if current route contains navPath
@@ -27,7 +23,6 @@ const TopBar = ({ topbarTitle, noBorderBottom }) => {
   const isCollapsed = useSelector((state) => state.sidebar.isCollapsed);
   const userInfo = useSelector((state) => state.login.userInfo);
   const [dropdownVisible, setDropdownVisible] = useState(false);
-  const cardCheckList = useSelector((state) => state.cardChecklist.cards);
   const [isLoading, setIsLoading] = useState(false);
   const togglingDrawer = useSelector((state) => state.sidebar.toggleDrawer);
 
@@ -42,32 +37,10 @@ const TopBar = ({ topbarTitle, noBorderBottom }) => {
     setDropdownVisible(!dropdownVisible);
   };
 
-  const handleLogout = async (event) => {
+  const handleLogout = (event) => {
     event.preventDefault();
-
     setIsLoading(true);
-
-    const action = await dispatch(
-      updateUserChecklists({
-        user_id: userInfo._id,
-        items: cardCheckList,
-      }),
-    );
-
-    // Check the status of the updateUserChecklists request
-    if (updateUserChecklists.rejected.match(action)) {
-      setIsLoading(false);
-      return;
-    }
-
-    localStorage.clear();
-    dispatch(resetStore());
-    dispatch(resetChartStore());
-    dispatch(clearIndividualPreferences());
-    dispatch(resetAllTasks());
-    dispatch(resetChecklist());
-    router.push('/account/login');
-
+    LogoutUser(dispatch, router);
     setIsLoading(false);
   };
